UserCourse modeli sisteme dahil edildi

+++ Presentation/Controllers/ReviewsController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Presentation.ActionFilters;
 using Services.Contracts;
+using System.Security.Claims;
 
 namespace Presentation.Controllers
 {
@@ -48,10 +49,12 @@ namespace Presentation.Controllers
 
         // Post
         [HttpPost]
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "User")]
         [ServiceFilter(typeof(ValidationFilterAttribute))]
         public async Task<IActionResult> CreateOneReviewAsync([FromBody] ReviewDtoForInsertion reviewDto)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
             //check courseId
             var course = await _manager
                 .CourseService
@@ -60,23 +63,33 @@ namespace Presentation.Controllers
             //check userId
             var user = await _manager
                 .AuthenticationService
-                .GetOneUserByIdAsync(reviewDto.UserId);
+                .GetOneUserByIdAsync(userId);
 
             if (course is null || user is null)
                 return NotFound();
 
             var review = await _manager
                 .ReviewService
-                .CreateOneReviewAsync(reviewDto);
+                .CreateOneReviewAsync(reviewDto, userId);
 
             return StatusCode(201, review);
         }
 
         // Delete
         [HttpDelete("{id:int}")]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public async Task<IActionResult> DeleteOneReviewAsync([FromRoute(Name = "id")] int id)
         {
+            var role = User.Claims
+               .FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            var review = await _manager.ReviewService.GetOneReviewByIdAsync(id, false);
+
+            if (review.UserId != userId && role == "User")
+                return BadRequest();
+
             await _manager.ReviewService.DeleteOneReviewAsync(id, false);
 
             return NoContent();
@@ -111,9 +124,18 @@ namespace Presentation.Controllers
 
         // Patch
         [HttpPatch("content/{id:int}")]
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "User")]
         public async Task<IActionResult> UpdateOneReviewContentAsync([FromRoute(Name = "id")] int id, [FromBody] string content)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            var review = await _manager
+                .ReviewService
+                .GetOneReviewByIdAsync(id, false);
+
+            if (review.UserId != userId)
+                return BadRequest();
+
             await _manager.ReviewService.UpdateOneReviewContentAsync(id, content, false);
 
             return NoContent();
